+++ engine/default-shaders/vertex.glsl
@@ -38,7 +38,7 @@ out vec3 lightPosAdj[LIGHT_COUNT];
 
 struct light
 {
-	int type;//0=empty (default),  1=ambient, 2=directional, 3=point, 4=spot
+	lowp int type;//0=empty (default),  1=ambient, 2=directional, 3=point, 4=spot
 	vec3 location,direction;//direction ignored if not spotlight; location ignored if ambient or directional
 	float angle;//spotlight only
 	float attenuation;//ignored on ambient
@@ -47,8 +47,8 @@ struct light
 	vec4 diffuseMultiply;//ignored on ambient
 	vec4 specularMultiply;//ignored on ambient
 	float shininess;//ignored on ambient
-	int negativeHandler; //0=no change (allow negatives), 1=clamp (min 0), 2=clamp negative (max 0), 3=absolute value
-	int negativeHandlerAlt; //same as negative handler but applies to specular only
+	lowp int negativeHandler; //0=no change (allow negatives), 1=clamp (min 0), 2=clamp negative (max 0), 3=absolute value
+	lowp int negativeHandlerAlt; //same as negative handler but applies to specular only
 };
 
 uniform light lights[LIGHT_COUNT];
